Updated the main class with a draw circle button
The button allows the user to click a position (the safe area near the center is probably the best spot) on the GUI where the program will autonomously generate points for a circle around the position of the mouse click. I am yet to test the pixel-size aspect of this but I know it was working when I was running my own code with a radius of 50, the only issue really being in relation to the size of it on the GUI being cut off due to the points of which are not safe to use with the Scara Arm.